Fix issues with Override components, refactored few pieces, and remove unnecessary imports
