+++ projects/angular-sdk-library/src/lib/_components/designSystemExtension/pulse/pulse.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit, Input, forwardRef } from '@angular/core';
-import { FeedContainerComponent } from '../../widget/feed-container/feed-container.component';
 import { ComponentMapperComponent } from '../../../_bridge/component-mapper/component-mapper.component';
 
 declare const window: any;
@@ -9,7 +8,7 @@ declare const window: any;
   templateUrl: './pulse.component.html',
   styleUrls: ['./pulse.component.scss'],
   standalone: true,
-  imports: [FeedContainerComponent, forwardRef(() => ComponentMapperComponent)]
+  imports: [forwardRef(() => ComponentMapperComponent)]
 })
 export class PulseComponent implements OnInit {
   @Input() pConn$: any;
