+++ projects/angular-sdk-library/src/lib/_components/field/user-reference/user-reference.component.ts
@@ -8,7 +8,6 @@ import { MatSelectModule } from '@angular/material/select';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { AngularPConnectService } from '../../../_bridge/angular-pconnect';
 import { Utils } from '../../../_helpers/utils';
-import { OperatorComponent } from '../../designSystemExtension/operator/operator.component';
 import { ComponentMapperComponent } from '../../../_bridge/component-mapper/component-mapper.component';
 
 declare const window: any;
@@ -24,7 +23,6 @@ const SEARCH_BOX = 'Search box';
   standalone: true,
   imports: [
     CommonModule,
-    OperatorComponent,
     ReactiveFormsModule,
     MatFormFieldModule,
     MatSelectModule,
