+++ projects/angular-sdk-library/src/lib/_components/infra/region/region.component.ts
@@ -2,7 +2,6 @@ import { Component, OnInit, Input, forwardRef, OnChanges, SimpleChanges } from '
 import { CommonModule } from '@angular/common';
 import { FormGroup } from '@angular/forms';
 import { ReferenceComponent } from '../reference/reference.component';
-// import { ComponentMapperComponent } from '../../../_bridge/component-mapper/component-mapper.component';
 import { ComponentMapperComponent } from '../../../_bridge/component-mapper/component-mapper.component';
 
 @Component({
