+++ projects/angular-sdk-library/src/lib/_components/template/one-column/one-column.component.ts
@@ -1,9 +1,6 @@
 import { Component, OnInit, Input, forwardRef, OnChanges, SimpleChanges } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormGroup } from '@angular/forms';
-import { CaseCreateStageComponent } from '../../designSystemExtension/case-create-stage/case-create-stage.component';
-import { ViewComponent } from '../../infra/view/view.component';
-import { RegionComponent } from '../../infra/region/region.component';
 import { ComponentMapperComponent } from '../../../_bridge/component-mapper/component-mapper.component';
 
 @Component({
@@ -11,7 +8,7 @@ import { ComponentMapperComponent } from '../../../_bridge/component-mapper/comp
   templateUrl: './one-column.component.html',
   styleUrls: ['./one-column.component.scss'],
   standalone: true,
-  imports: [CommonModule, ViewComponent, CaseCreateStageComponent, RegionComponent, forwardRef(() => ComponentMapperComponent)]
+  imports: [CommonModule, forwardRef(() => ComponentMapperComponent)]
 })
 export class OneColumnComponent implements OnInit, OnChanges {
   @Input() pConn$: any;
