+++ projects/angular-sdk-library/src/lib/_components/template/three-column-page/three-column-page.component.ts
@@ -1,7 +1,5 @@
 import { Component, OnInit, Input, forwardRef } from '@angular/core';
 import { FormGroup } from '@angular/forms';
-// import { ComponentMapperComponent } from '../../../_bridge/component-mapper/component-mapper.component';
-import { ThreeColumnComponent } from '../three-column/three-column.component';
 import { ComponentMapperComponent } from '../../../_bridge/component-mapper/component-mapper.component';
 
 
@@ -10,7 +8,7 @@ import { ComponentMapperComponent } from '../../../_bridge/component-mapper/comp
   templateUrl: './three-column-page.component.html',
   styleUrls: ['./three-column-page.component.scss'],
   standalone: true,
-  imports: [ThreeColumnComponent, forwardRef(() => ComponentMapperComponent)]
+  imports: [forwardRef(() => ComponentMapperComponent)]
 })
 export class ThreeColumnPageComponent implements OnInit {
   @Input() pConn$: any;
