+++ projects/angular-sdk-library/src/lib/_helpers/common.ts
