+++ projects/angular-sdk-library/src/public-api.ts
@@ -129,9 +129,12 @@ export * from './lib/_services/endpoints';
 
 export * from './lib/_helpers/case-utils';
 export * from './lib/_helpers/currency-utils';
+export * from './lib/_helpers/common';
 export * from './lib/_helpers/date-format-utils';
 export * from './lib/_helpers/event-util';
 export * from './lib/_helpers/field-group-utils';
+export * from './lib/_helpers/formatters/format-utils';
+export * from './lib/_helpers/formatters/currency-map';
 export * from './lib/_helpers/tab-utils';
 export * from './lib/_helpers/template-utils';
 export * from './lib/_helpers/utils';
